[package] more reliable method for determining standard library-ness
We implicitly extern standard library modules. Our method of determining
whether a module is in the standard library is a little unreliable. In
particular, I'm seeing lots of flaky errors on windows/mac CI when I
start doing more complicated packaging tests.

I looked into the best ways to do this, turns out there's no reliable
way, so tools that need to do this generally just parse the Python docs
for a listing and save it. I took `isort`'s lists and called it a day.

ghstack-source-id: 7ff285926c26b87b6aff2c2e721eebbf05feb065
Pull Request resolved: https://github.com/pytorch/pytorch/pull/51694